Moved the GeofenceBroadcastReceiver to a separate class.
